Move StyleViews to Own File
