+++ Sources/CodeEditSourceEditor/Controller/TextViewController.swift
@@ -271,7 +271,7 @@ public class TextViewController: NSViewController {
     // MARK: Paragraph Style
 
     /// A default `NSParagraphStyle` with a set `lineHeight`
-    internal lazy var paragraphStyle: NSMutableParagraphStyle = generateParagraphStyle()
+    package lazy var paragraphStyle: NSMutableParagraphStyle = generateParagraphStyle()
 
     private func generateParagraphStyle() -> NSMutableParagraphStyle {
         // swiftlint:disable:next force_cast
@@ -292,59 +292,6 @@ public class TextViewController: NSViewController {
         styleGutterView()
 
         highlighter?.invalidate()
-    }
-
-    /// Style the text view.
-    package func styleTextView() {
-        textView.selectionManager.selectionBackgroundColor = theme.selection
-        textView.selectionManager.selectedLineBackgroundColor = getThemeBackground()
-        textView.selectionManager.highlightSelectedLine = isEditable
-        textView.selectionManager.insertionPointColor = theme.insertionPoint
-        paragraphStyle = generateParagraphStyle()
-        textView.typingAttributes = attributesFor(nil)
-    }
-
-    /// Finds the preferred use theme background.
-    /// - Returns: The background color to use.
-    private func getThemeBackground() -> NSColor {
-        if useThemeBackground {
-            return theme.lineHighlight
-        }
-
-        if systemAppearance == .darkAqua {
-            return NSColor.quaternaryLabelColor
-        }
-
-        return NSColor.selectedTextBackgroundColor.withSystemEffect(.disabled)
-    }
-
-    /// Style the gutter view.
-    package func styleGutterView() {
-        gutterView.frame.origin.y = -scrollView.contentInsets.top
-        gutterView.selectedLineColor = useThemeBackground ? theme.lineHighlight : systemAppearance == .darkAqua
-        ? NSColor.quaternaryLabelColor
-        : NSColor.selectedTextBackgroundColor.withSystemEffect(.disabled)
-        gutterView.highlightSelectedLines = isEditable
-        gutterView.font = font.rulerFont
-        gutterView.backgroundColor = useThemeBackground ? theme.background : .textBackgroundColor
-        if self.isEditable == false {
-            gutterView.selectedLineTextColor = nil
-            gutterView.selectedLineColor = .clear
-        }
-    }
-
-    /// Style the scroll view.
-    package func styleScrollView() {
-        guard let scrollView = view as? NSScrollView else { return }
-        scrollView.drawsBackground = useThemeBackground
-        scrollView.backgroundColor = useThemeBackground ? theme.background : .clear
-        if let contentInsets {
-            scrollView.automaticallyAdjustsContentInsets = false
-            scrollView.contentInsets = contentInsets
-        } else {
-            scrollView.automaticallyAdjustsContentInsets = true
-        }
-        scrollView.contentInsets.bottom = (contentInsets?.bottom ?? 0) + bottomContentInsets
     }
 
     deinit {
